+++ src/tags/tags.service.ts
@@ -1,9 +1,11 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { ForbiddenException, Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Tag } from './entities/tags.entity';
 import { Repository } from 'typeorm';
 import { addTagDTO } from './dto/tags.dto';
 import { ConfirmationMsg } from 'src/utils/confirmation.interface';
+import { UserWithoutPassword } from 'src/auth/interfaces/user.interface';
+import { ProductsService } from 'src/products/products.service';
 
 @Injectable()
 export class TagsService {
@@ -16,9 +18,10 @@ export class TagsService {
         return await this.tagRepository.find();
     }
 
-    async createTag(body: addTagDTO): Promise<ConfirmationMsg> {
+    async createTag(body: addTagDTO, user: UserWithoutPassword): Promise<ConfirmationMsg> {
         const tag = await this.tagRepository.save({
-            name: body.name
+            name: body.name,
+            user,
         })
         return {
             id: tag.id,
@@ -38,8 +41,11 @@ export class TagsService {
         return tag;
     }
 
-    async deleteTag(id: string): Promise<ConfirmationMsg> {
+    async deleteTag(id: string, user: UserWithoutPassword): Promise<ConfirmationMsg> {
         const tag: Tag = await this.findTagById(id);
+        if (tag.user.id !== user.id) {
+            throw new ForbiddenException("Forbidden!");
+        }
         const tagId = tag.id;
         await this.tagRepository.remove(tag);
         return {
